Counting jobs in progress for Resque
The Macro is only counting the jobs in the queues and skips 
counting those that are being actively worked on. We need to look
at the workers to see which jobs are active, otherwise we could have
the case where we spin up n workers for n jobs, each worker takes a
job, there are 0 jobs left on the queue, and all the workers are spun
down even though they are in the process of working.