Introduce HireFire::Macro::Resque::Cache
This change adds a caching mechanism to the Resque macro.

Since we have to scan the entire delayed set in Redis to acquire the sizes of the requested queues
containing jobs scheduled to run now, we can instead process the size of every queue associated with
the scanned jobs, whether requested or not, and cache it for a certain amount of time.

This approach allows us to avoid scanning the entire delayed set on subsequent requests within the
cache duration. Instead, we can access and aggregate the cached data. This will significantly
improve throughput for applications that have a large number of scheduled jobs and/or a large number
of Resque process types that need to be monitored.